requirements.txt
fastapi[standard]
uvicorn[standard]
python-dotenv
sqlalchemy
alembic
pydantic
bcrypt
passlib[bcrypt]
python-jose[cryptography]
google-generativeai
langchain
langchain-community
langchain-huggingface
faiss-cpu
sentence-transformers
transformers
spacy
huggingface-hub
PyPDF2
python-multipart
regex
httpx
langchain-core
python-docx
openpyxl
pillow
pandas
numpy
tiktoken
